feat: convert formfield demo to playground
